Reorganize flat Finance module into domain modules (1.2.0)
Split the single Finance module into cohesive domain modules with no change to
any function's signature or result:

- Finance.CashFlow  — npv, xnpv, irr, xirr, mirr
- Finance.TVM       — pv, fv, pmt, nper, rate
- Finance.Depreciation — sln, syd, ddb, db
- Finance.Returns   — volatility
- Finance.Shared    — internal helpers + options schema (@moduledoc false)

Extract the root-finder into a Finance.Solver behaviour with a default
Finance.Solver.Newton implementation, selectable per-call via the :solver
option or globally via `config :finance, solver: MySolver` — the seam for a
future Nx/GPU solver. Newton.solve/2 now uses a single `with` instead of two
nested `case`s.

Finance becomes a facade: every previously-public function is an
@deprecated defdelegate to its new home (removed in 2.0). Public types live in
Finance so ExDoc doesn't reference the hidden Finance.Shared.

Tests split by module (doctests repointed), plus a data-driven test asserting
every delegator matches its new home and a test proving the :solver seam.

All green: 135 tests (24 doctests, 3 properties, 108 unit), credo --strict,
dialyzer, mix format, 100% coverage, 0 doc warnings.
